Implemented Handle and HandleFunc with tests. Reorganized for improve referencing.
