+++ router_test.go
@@ -9,6 +9,7 @@
 package vestigo
 
 import (
+	"bytes"
 	"fmt"
 	"net/http"
 	"net/http/httptest"
@@ -532,6 +533,138 @@ func TestRouterAddInvalidMethod(t *testing.T) {
 	assert.Panics(t, func() {
 		r.Add("INVALID", "/", func(w http.ResponseWriter, req *http.Request) {})
 	})
+}
+
+func TestMethodSpecificAddRoute(t *testing.T) {
+	router := NewRouter()
+	m := map[string]func(path string, handler http.HandlerFunc){
+		"GET":     router.Get,
+		"POST":    router.Post,
+		"CONNECT": router.Connect,
+		"DELETE":  router.Delete,
+		"PATCH":   router.Patch,
+		"PUT":     router.Put,
+		"TRACE":   router.Trace,
+	}
+	f := func(w http.ResponseWriter, r *http.Request) {
+		w.WriteHeader(200)
+		w.Write([]byte("success-" + r.Method))
+	}
+	path := "/test"
+	for _, v := range m {
+		v(path, f)
+	}
+	for k := range m {
+		w := httptest.NewRecorder()
+		r, err := http.NewRequest(k, path, nil)
+		if err != nil {
+			t.Errorf("Failed to create a new request, method: %s, path: %s", k, path)
+		}
+		router.ServeHTTP(w, r)
+		if w.Code != 200 || w.Body.String() != "success-"+k {
+			t.Errorf("Invalid response, method: %s, path: %s, code: %s, body: %s", k, path, w.Code, w.Body.String())
+		}
+	}
+
+}
+
+func TestTrace(t *testing.T) {
+	router := NewRouter()
+	AllowTrace = true
+	defer func() {
+		AllowTrace = false
+	}()
+	path := "/test"
+	router.Get(path, func(w http.ResponseWriter, r *http.Request) {
+		w.WriteHeader(200)
+		w.Write([]byte(""))
+	})
+	router.Patch(path+"/split", func(w http.ResponseWriter, r *http.Request) {
+		w.WriteHeader(200)
+		w.Write([]byte(""))
+	})
+	router.Connect(path+"/split/again", func(w http.ResponseWriter, r *http.Request) {
+		w.WriteHeader(200)
+		w.Write([]byte(""))
+	})
+
+	w := httptest.NewRecorder()
+	r, err := http.NewRequest("TRACE", path+"/split", bytes.NewBufferString("awesome trace"))
+	if err != nil {
+		t.Errorf("Failed to create a new request, method: %s, path: %s", "TRACE", path)
+	}
+	router.ServeHTTP(w, r)
+	if w.Code != 200 || w.Body.String() != "awesome trace" || w.Header().Get("Content-Type") != "message/http" {
+		t.Errorf("Invalid TRACE response, method: %s, path: %s, code: %s, body: %s", "TRACE", path, w.Code, w.Body.String())
+	}
+}
+
+func TestHead(t *testing.T) {
+	router := NewRouter()
+	path := "/test"
+	router.Get(path, func(w http.ResponseWriter, r *http.Request) {
+		w.Header().Add("X-TestHeader", "true")
+		w.WriteHeader(200)
+		w.Write([]byte("some return body"))
+	})
+
+	w := httptest.NewRecorder()
+	r, err := http.NewRequest("HEAD", path, nil)
+	if err != nil {
+		t.Errorf("Failed to create a new request, method: %s, path: %s", "HEAD", path)
+	}
+	router.ServeHTTP(w, r)
+	if w.Code != 200 || w.Body.String() != "" || w.Header().Get("X-TestHeader") != "true" {
+		t.Errorf("Invalid HEAD response, method: %s, path: %s, code: %s, body: %s", "HEAD", path, w.Code, w.Body.String())
+	}
+}
+
+func TestHandleAddRoute(t *testing.T) {
+	router := NewRouter()
+	f := func(w http.ResponseWriter, r *http.Request) {
+		w.WriteHeader(200)
+		w.Write([]byte("success-" + r.Method))
+	}
+	path := "/test"
+	router.Handle(path, http.HandlerFunc(f))
+	for _, v := range methods {
+		if v == "HEAD" || v == "OPTIONS" {
+			continue
+		}
+		w := httptest.NewRecorder()
+		r, err := http.NewRequest(v, path, nil)
+		if err != nil {
+			t.Errorf("Failed to create a new request, method: %s, path: %s", v, path)
+		}
+		router.ServeHTTP(w, r)
+		if w.Code != 200 || w.Body.String() != "success-"+v {
+			t.Errorf("Invalid response, method: %s, path: %s, code: %s, body: %s", v, path, w.Code, w.Body.String())
+		}
+	}
+}
+
+func TestHandleFuncAddRoute(t *testing.T) {
+	router := NewRouter()
+	f := func(w http.ResponseWriter, r *http.Request) {
+		w.WriteHeader(200)
+		w.Write([]byte("success-" + r.Method))
+	}
+	path := "/test"
+	router.HandleFunc(path, f)
+	for _, v := range methods {
+		if v == "HEAD" || v == "OPTIONS" {
+			continue
+		}
+		w := httptest.NewRecorder()
+		r, err := http.NewRequest(v, path, nil)
+		if err != nil {
+			t.Errorf("Failed to create a new request, method: %s, path: %s", v, path)
+		}
+		router.ServeHTTP(w, r)
+		if w.Code != 200 || w.Body.String() != "success-"+v {
+			t.Errorf("Invalid response, method: %s, path: %s, code: %s, body: %s", v, path, w.Code, w.Body.String())
+		}
+	}
 }
 
 func TestRouterServeHTTP(t *testing.T) {
